Move UI schema components from API into SDK viavai package
