--- api/src/ui/__init__.py
@@ -1,1 +1,0 @@
-from .page import Page
