+++ sdk/viavai/ui/__init__.py
@@ -1,36 +1,1 @@
-# https://ui.shadcn.com/docs/components/radix/avatar
-# https://ui.shadcn.com/docs/components/radix/badge
-# https://ui.shadcn.com/docs/components/radix/button
-# https://ui.shadcn.com/docs/components/radix/button-group
-# https://ui.shadcn.com/docs/components/radix/card
-# https://ui.shadcn.com/docs/components/radix/checkbox
-# https://ui.shadcn.com/docs/components/base/switch
-# https://ui.shadcn.com/docs/components/radix/data-table
-# https://ui.shadcn.com/docs/components/radix/date-picker
-
-# https://ui.shadcn.com/docs/components/base/select
-
-# https://ui.shadcn.com/docs/components/base/field
-# https://ui.shadcn.com/docs/components/radix/dialog -> With form
-# https://ui.shadcn.com/docs/components/radix/empty
-# https://ui.shadcn.com/docs/components/base/hover-card
-# https://ui.shadcn.com/docs/components/base/input
-# https://ui.shadcn.com/docs/components/base/input-group
-
-# https://ui.shadcn.com/docs/components/base/item
-# https://ui.shadcn.com/docs/components/base/label
-# https://ui.shadcn.com/docs/components/base/pagination
-# https://ui.shadcn.com/docs/components/base/radio-group
-# https://ui.shadcn.com/docs/components/base/separator
-
-
-# Overall
-# https://ui.shadcn.com/docs/components/base/skeleton
-# 
-# https://ui.shadcn.com/docs/components/base/table
-# https://ui.shadcn.com/docs/components/base/tabs
-# https://ui.shadcn.com/docs/components/base/textarea
-# https://ui.shadcn.com/docs/components/base/typography
-
-
 from .page import Page
