+++ sdk/viavai/ui/hero.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from src.ui.__root__ import Component
+from .__root__ import Component
 
 
 @dataclass
