+++ sdk/viavai/ui/tabs.py
@@ -1,21 +1,25 @@
 from dataclasses import dataclass, field
 
-from src.ui.__root__ import Component
-from src.ui.tabs import Tab, Tabs
-from src.ui.button import Button, ButtonVariants
-from src.ui.columns import Column, Columns
-from src.ui.hero import Hero
-from src.ui.input import Input
-from src.ui.menu import Menu
-from src.ui.separator import Separator
-from src.ui.table import Table
+from .__root__ import Component
+from .button import Button, ButtonVariants
+from .columns import Column, Columns
+from .hero import Hero
+from .input import Input
+from .separator import Separator
+from .table import Table
 
 
 @dataclass
-class Page:
-    """Top-level component that represents a page in the UI."""
-
+class Tab(Component):
+    name: str
     _children: list[Component] = field(default_factory=list)
+
+    def __iter__(self):
+        yield 'type', 'tab'
+        yield 'props', {
+            'name': self.name,
+        }
+        yield 'children', [dict(child) for child in self._children]
 
     def hero(self, title: str, subtitle: str | None = None) -> Hero:
         hero = Hero(title=title, subtitle=subtitle)
@@ -37,16 +41,10 @@ class Page:
         self._children.append(columns)
         return [columns.column(width=width) for width in widths]
 
-    def menu(self) -> Menu:
-        menu = Menu()
-        self._children.append(menu)
-        return menu
-
-    def tabs(self, names: list[str]) -> list[Tab]:
-        tabs = Tabs()
-        self._children.append(tabs)
-        return [tabs.tab(name=name) for name in names]
-
+    def separator(self) -> Separator:
+        separator = Separator()
+        self._children.append(separator)
+        return separator
 
     def input(
         self,
@@ -64,11 +62,21 @@ class Page:
         self._children.append(input_component)
         return input_component
 
-    def separator(self) -> Separator:
-        separator = Separator()
-        self._children.append(separator)
-        return separator
+
+@dataclass
+class Tabs(Component):
+    _tabs: list[str] = field(default_factory=list)
+    _children: list[Tab] = field(default_factory=list)
+
+    def tab(self, name: str) -> Tab:
+        tab = Tab(name=name)
+        self._tabs.append(name)
+        self._children.append(tab)
+        return tab
 
     def __iter__(self):
-        for component in self._children:
-            yield dict(component)
+        yield 'type', 'tabs'
+        yield 'props', {
+            'tabs': self._tabs,
+        }
+        yield 'children', [dict(tab) for tab in self._children]
